+++ src/sider/ftrl/saksbehandling/stegLister/stegListeIkkeYrkesaktivFlyt.tsx
@@ -2,7 +2,7 @@ import { FANE_STATUS } from "../../../../felleskomponenter/stegvelger";
 import { VurderingInngang } from "../stegKomponenter/vurderingInngang/vurderingInngang";
 import { VurderingPerioder } from "../stegKomponenter/vurderingPeriode/vurderingPerioder";
 import { VurderingVedtak } from "../stegKomponenter/vurderingVedtak/vurderingVedtak";
-import { VurderingBestemmelserV2 } from "../stegKomponenter/vurderingBestemmelse/vurderingBestemmelseNY";
+import { VurderingBestemmelse } from "../stegKomponenter/vurderingBestemmelse/vurderingBestemmelse";
 
 const initialInngangSteg = {
   id: "Inngang",
@@ -21,7 +21,7 @@ const initialBestemmelseSteg = {
   status: FANE_STATUS.UBEHANDLET,
   aktivtSteg: false,
   vedtakSteg: false,
-  komponent: VurderingBestemmelserV2,
+  komponent: VurderingBestemmelse,
 };
 
 const initialPeriodeSteg = {
